requirements.txt
python-dotenv>=1.0.0
requests>=2.28.2
tqdm>=4.65.0